型変更

+++ sql/4.6/orcadbup_tensuplus_4.6.0-1.sql
@@ -9,6 +9,6 @@
 
 \set ON_ERROR_STOP
 
-alter table tbl_tensuplus add column jgncnt1w numeric(3,0);
+alter table tbl_tensuplus add column jgncnt1w smallint;
 alter table tbl_tensuplus alter jgncnt1w set default 0;
 update tbl_tensuplus set jgncnt1w = 0;
